feat: refactor blog details page to fetch blog data
